Allocations/Budgets: rm almost all props
This moves us closer to being able to implement proper routing, where
the ideal ReactRouter layout prevents passing any props from App.js to
Budgets.

Note that this was a pretty clean restructuring – almost all of the
complexity in App.js was there to support Budgets, and could be easily
moved to Budgets.js instead.
