+++ apps/allocations/app/components/App/App.js
@@ -4,15 +4,12 @@ import { Button, Header, IconPlus, Main, SidePanel, SyncIndicator } from '@arago
 
 import { IdentityProvider } from '../LocalIdentityBadge/IdentityManager'
 import { Empty } from '../Card'
-import { NewAllocation, NewBudget } from '../Panel'
+import { NewBudget } from '../Panel'
 import { AllocationsHistory, Budgets } from '.'
-import { Deactivate } from '../Modal'
 
 const App = () => {
   const [ panel, setPanelRaw ] = useState(null)
   const [ panelOpen, setPanelOpen ] = useState(false)
-  const [ isModalVisible, setModalVisible ] = useState(false)
-  const [ currentBudgetId, setCurrentBudgetId ] = useState('')
   const { api, appState } = useAragonApi()
   const { allocations = [], budgets = [], isSyncing = true } = appState
 
@@ -26,49 +23,16 @@ const App = () => {
     }
   }
 
-  const saveBudget = ({ id, amount, name, token }) => {
-    if (id) {
-      api.setBudget(id, amount, name).toPromise()
-    } else {
-      api
-        .newAccount(
-          name,             // _metadata
-          token.address,    // _token
-          true,             // hasBudget
-          amount
-        )
-        .toPromise()
-    }
+  const saveBudget = ({ amount, name, token }) => {
+    api
+      .newAccount(
+        name,             // _metadata
+        token.address,    // _token
+        true,             // hasBudget
+        amount
+      )
+      .toPromise()
     setPanel(null)
-  }
-
-  const onSubmitAllocation = ({
-    addresses,
-    description,
-    budgetId,
-    period = 0,
-    balance,
-  }) => {
-    const emptyIntArray = new Array(addresses.length).fill(0)
-    api.setDistribution(
-      addresses,
-      emptyIntArray, // unused
-      emptyIntArray, // unused
-      '', // unused
-      description,
-      emptyIntArray, // unused
-      emptyIntArray, // unused
-      budgetId, // account or allocation id...budgetId
-      '1', // recurrences, 1 for now
-      Math.floor(new Date().getTime()/1000), // startTime, now for now
-      period,
-      balance, // amount
-    ).toPromise()
-    setPanel(null)
-  }
-
-  const onSubmitDeactivate = () => { // TODO id => {
-    closeModal()
   }
 
   // TODO: Fix this
@@ -82,46 +46,6 @@ const App = () => {
       content: NewBudget,
       data: { heading: 'New budget', saveBudget },
     })
-  }
-
-  const onNewAllocation = (budgetId) => {
-    const { balances } = appState
-    setPanel({
-      content: NewAllocation,
-      data: {
-        budgetId,
-        heading: 'New allocation',
-        onSubmitAllocation,
-        budgets,
-        balances,
-      },
-    })
-  }
-
-  const onEdit = id => {
-    const editingBudget = budgets.find(budget => budget.id === id)
-    setPanel({
-      content: NewBudget,
-      data: {
-        heading: 'Edit budget',
-        saveBudget,
-        editingBudget,
-      },
-    })
-  }
-
-  const onDeactivate = id => {
-    setModalVisible(true)
-    setCurrentBudgetId(id)
-  }
-
-  const onReactivate = () => { // TODO id => {
-    //api.reactivateBudget(id)
-  }
-
-  const closeModal = () => {
-    setModalVisible(false)
-    setCurrentBudgetId('')
   }
 
   const handleResolveLocalIdentity = address =>
@@ -153,24 +77,12 @@ const App = () => {
                   />
                 }
               />
-              <Budgets
-                budgets={budgets}
-                onNewAllocation={onNewAllocation}
-                onEdit={onEdit}
-                onDeactivate={onDeactivate}
-                onReactivate={onReactivate}
-              />
+              <Budgets setPanel={setPanel} />
               <SyncIndicator visible={isSyncing} />
             </React.Fragment>
           )
         }
         { !!allocations.length && <AllocationsHistory allocations={allocations} /> }
-        <Deactivate
-          visible={isModalVisible}
-          budgetId={currentBudgetId}
-          onClose={closeModal}
-          onSubmit={onSubmitDeactivate}
-        />
         <SidePanel
           title={(panel && panel.data.heading) || ''}
           opened={panelOpen}
